feste länge

+++ MI-Medienpool_DDL.sql
@@ -61,7 +61,7 @@ zubehoer VARCHAR2(300) DEFAULT 'Kein Zubehoer verfuegbar' NOT NULL
 );
 
 CREATE TABLE benutzer (
-benutzer_matrnr INTEGER,
+benutzer_matrnr NUMBER(5),
 benutzer_name VARCHAR2(60),
 benutzer_verstosspunkte INTEGER DEFAULT 0,
 gesperrt VARCHAR2(4) DEFAULT 'Nein'
@@ -117,7 +117,6 @@ ALTER TABLE geraetearten ADD CONSTRAINT geraeteart_id_check CHECK (geraeteart_id
 ALTER TABLE geraetearten ADD CONSTRAINT geraeteart_check CHECK  (geraeteart IN ('Laptop', 'Wearable', 'Smartphone', 'Tablet', 'Sonstiges'));
 ALTER TABLE verstossarten ADD CONSTRAINT verstossarten_id_check CHECK (verstossart_id IN (1 ,2 ,3));
 ALTER TABLE verstossarten ADD CONSTRAINT verstossart_beschreibung_check CHECK (verstossart_beschreibung in ('Leihdauer ueberzogen', 'Geraet defekt zurueckgebracht', 'Geraet verkratzt/beschaedigt'));
-
 -- Primary Key's
 ALTER TABLE benutzer ADD CONSTRAINT benutzer_matrnr_pk PRIMARY KEY (benutzer_matrnr);
 ALTER TABLE geraete_verliehen ADD CONSTRAINT verleih_id_pk PRIMARY KEY (verleih_id);
